requirements.txt

streamlit
pandas
streamlit-authenticator
pyyaml
streamlit-calendar
